Split grpc actor tests in a separate file.
